mysite/pyproject.toml
[project]
name = "mysite"
version = "0.1.0"
description = ""
authors = [
    {name = "Gevor",email = "gev.srg.044@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django==5.2.1",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "django-debug-toolbar==5.2.0",
    "django-filter==25.1",
    "djangorestframework==3.16.0",
    "docutils==0.21.2",
    "drf-spectacular==0.28.0",
    "inflection==0.5.1",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "pillow==11.2.1",
    "pytz==2022.7.1",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "rpds-py==0.25.1",
    "sqlparse==0.5.3",
    "typing_extensions==4.13.2",
    "tzdata==2025.2",
    "uritemplate==4.1.1"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
